날씨 리스트 레이아웃 작업

+++ app/src/main/java/com/bentley/localweather/data/mapper/WeatherMapper.kt
@@ -27,7 +27,7 @@ class ConsolidatedWeatherMapper @Inject constructor() :
                     weatherStateAbbr,
                     date.getDate(),
                     temp.toInt(),
-                    humidity
+                    humidity.toInt()
                 )
             }
         }[0]
