+++ app/src/main/java/com/bentley/localweather/domain/entity/Weather.kt
@@ -15,5 +15,5 @@ data class ConsolidatedWeather(
     val weatherStateAbbr: String,
     val date: Int,
     val temp: Int,
-    val humidity: Float,
+    val humidity: Int,
 )
